+++ src/test/java/racingcar/RacingCarGameTest.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.dto.RacingCar;
-import racingcar.utils.RacingCarGame;
-import racingcar.utils.RacingCarGameOutput;
+import racingcar.domain.RacingCarGame;
+import racingcar.domain.RacingCarGameOutput;
 import utils.OutputUtil;
 
 public class RacingCarGameTest {
